Flatten app package
Flatten the app package rather than having one domain per package. This
is a better fit for a demonstration or application template. The package
structure of an application should start simply, with few packages and
many well named files. It should only be decomposed once the structure
and flow of the application has become established.

This commit also renames the app package to controller. This better
reflects the role of this package as the coordinating business logic of
the application.